Give cb_position the width every other struct already has
cb_position carried a uint32_t spine and page while every struct that
came after it — cb_session_event, cb_toc_entry, cb_search_hit, the
locator calls — speaks size_t. The split was harmless in range (no book
has four billion units) but every binding had to carry a cast at exactly
one struct, and one abi.rs test carried a comment explaining it. Both
fields are size_t now.

This is a Contract-tier ABI break, taken because the workspace is 0.x
and the wart compounds with every new binding written against it. The
.NET public surface had mirrored the same split (Position was uint while
Locator, SessionEvent and GoToAnchor are int), so it closes there too:
Position(int, int), transcription nuint, casts in the tests gone. Swift
compiles unchanged — size_t imports as Int, which is what its Position
already holds. Header regenerated; NDK clang compiles it for both ABIs.

+++ crates/chapbook-ffi/src/session.rs
@@ -112,8 +112,8 @@ pub enum cb_book_kind {
 #[repr(C)]
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct cb_position {
-    pub spine: u32,
-    pub page: u32,
+    pub spine: usize,
+    pub page: usize,
 }
 
 /// The page box, in logical units, plus the scale that turns it into
@@ -555,8 +555,8 @@ pub unsafe extern "C" fn cb_session_position(
         out!(
             position,
             cb_position {
-                spine: at.spine as u32,
-                page: at.page as u32,
+                spine: at.spine,
+                page: at.page,
             },
             "position"
         );
